+++ admin/application/controllers/Menu.php
@@ -19,15 +19,51 @@ class Menu extends CI_Controller
 
     public function index()
     {
+        if(!permission("menu", "show")){
+            redirect(base_url());
+        }
+
         $this->breadcrumbs->unshift('Anasayfa', '/');
         $this->breadcrumbs->push('Menü İşlemleri', '/menu');
 
         $viewData = new stdClass();
+        /* Pagination Start */
+        $config["base_url"] = base_url("$this->tableName/index");
+        $config["total_rows"] = $this->menu_model->get_count();
+        $config["uri_segment"] = 3;
+        $config["per_page"] = 10;
+        $config["num_links"] = 2;
+
+        $config['full_tag_open'] = "<nav class='search-results-navigation'> <ul class='pagination'>";
+        $config['full_tag_close'] = '</ul></nav>';
+        $config['num_tag_open'] = '<li>';
+        $config['num_tag_close'] = '</li>';
+        $config['cur_tag_open'] = '<li class="active"><a href="#">';
+        $config['cur_tag_close'] = '</a></li>';
+        $config['prev_tag_open'] = '<li>';
+        $config['prev_tag_close'] = '</li>';
+        $config['first_tag_open'] = '<li>';
+        $config['first_tag_close'] = '</li>';
+        $config['last_tag_open'] = '<li>';
+        $config['last_tag_close'] = '</li>';
+        $config['prev_link'] = '<i class="fa fa-long-arrow-left"></i>Geri';
+        $config['prev_tag_open'] = '<li>';
+        $config['prev_tag_close'] = '</li>';
+        $config['next_link'] = 'İleri<i class="fa fa-long-arrow-right"></i>';
+        $config['next_tag_open'] = '<li>';
+        $config['next_tag_close'] = '</li>';
+
+        $this->pagination->initialize($config);
+        $page = ($this->uri->segment(3)) ? $this->uri->segment(3): 0;
+        $viewData->links = $this->pagination->create_links();
+        /* Pagination End */
 
         //Tablodan verilerin çekilmesi.
-        $items = $this->menu_model->getAll(
+        $items = $this->menu_model->get_records(
             array("isSubmenu" => 0),
-            "rank ASC"
+            "rank ASC",
+            $config["per_page"],
+            $page
         );
 
         //View'e gönderilen verilerin set edilmesi.
@@ -41,6 +77,10 @@ class Menu extends CI_Controller
     }
 
     public function addForm(){
+
+        if(!permission("menu", "add")){
+            redirect(base_url());
+        }
 
         $items = $this->menu_model->getAll(
             array("isSubmenu" => 0),
@@ -65,6 +105,10 @@ class Menu extends CI_Controller
 
     public function addItem()
     {
+        if(!permission("menu", "add")){
+            redirect(base_url());
+        }
+
         $items = $this->menu_model->getAll(
             array("isSubmenu" => 0),
             "rank ASC"
@@ -164,6 +208,10 @@ class Menu extends CI_Controller
 
     public function updateForm($id){
 
+        if(!permission("menu", "edit")){
+            redirect(base_url());
+        }
+
         $this->breadcrumbs->unshift('Anasayfa', '/', false);
         $this->breadcrumbs->push('Menü İşlemleri', '/menu');
         $this->breadcrumbs->push('Menü Düzenle','/');
@@ -196,6 +244,10 @@ class Menu extends CI_Controller
 
     public function updateItem($id)
     {
+        if(!permission("menu", "edit")){
+            redirect(base_url());
+        }
+
         $item = $this->menu_model->get(
             array(
                 "id" => $id
@@ -221,6 +273,7 @@ class Menu extends CI_Controller
             $data['title'] = $this->input->post('title');
             $data['content'] = htmlspecialchars($this->input->post('content'));
             $data['isSubmenu'] = $this->input->post('isSubmenu');
+            $data['isFooter'] = $this->input->post('isFooter');
             $data['url'] = permalink($this->input->post('url'));
             $data['seo'] = json_encode($this->input->post('seo'), JSON_UNESCAPED_UNICODE);
 
@@ -322,6 +375,10 @@ class Menu extends CI_Controller
 
     public function deleteItem($id)
     {
+        if(!permission("menu", "delete")){
+            redirect(base_url());
+        }
+
         $getItem = $this->menu_model->get(array("id" => $id));
 
         $delete = $this->menu_model->delete(
